skills/skill-toon-openapi/requirements.txt
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
coverage==7.10.7
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==26.0
pathable==0.4.4
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.2
pytest-cov==7.1.0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rpds-py==0.27.1
six==1.17.0
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
